Generalise the Data.Machine.Group module
I wanted to stream some data to sequentially named files without having to load
everything for each file into memory at a time.  This patch lets us group inputs
with a parameter, like a counter, which makes it easy to stream groups of data
that is related.

The old functions I've changed to have a `foo_` suffix that fits existing
conventions in this new Data.Machine.Group' module.